NAS-115593: Improvements after review, part 2
