+++ src/app/modules/ix-forms/components/ix-dynamic-form/ix-dynamic-form-item/ix-dynamic-form-item.component.scss
@@ -3,3 +3,7 @@
   opacity: 0.38;
   pointer-events: none;
 }
+
+.indent {
+  padding-left: 16px;
+}
