--- src/app/pages/applications/forms/chart-wizard/chart-wizard.component.html
@@ -1,22 +1,0 @@
-<ix-modal-header [title]="title" [loading]="isLoading"></ix-modal-header>
-<mat-card>
-  <mat-card-content>
-    <form [formGroup]="form" class="ix-form-container" (submit)="onSubmit()">
-      <ix-dynamic-form
-        [dynamicSection]="dynamicSection"
-        [dynamicForm]="form"
-        (addListItem)="addFormListItem($event)"
-        (deleteListItem)="deleteFormListItem($event)"
-      ></ix-dynamic-form>
-
-      <div class="form-actions">
-        <button
-          mat-button
-          type="submit"
-          [disabled]="form.invalid || isLoading"
-          color="primary"
-        >{{ 'Save' | translate }}</button>
-      </div>
-    </form>
-  </mat-card-content>
-</mat-card>
