--- src/app/pages/applications/forms/chart-wizard/chart-wizard.component.scss
@@ -1,3 +1,0 @@
-.form-actions {
-  margin: 8px 10px;
-}
